return a result instead of balance, make project generic over T, add deposit_id to project
