+++ pallets/deposits/Cargo.toml
@@ -33,6 +33,7 @@ orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-li
 common-types = { path = "../../libs/common-types" }
 orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library", branch = "polkadot-v0.9.39" }
 pallet-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.39"}
+common-runtime = { path = "../../runtime/common"}
 
 [features]
 default = ["std"]
@@ -44,6 +45,7 @@ std = [
 	"scale-info/std",
 	"sp-runtime/std",
 	"pallet-xcm/std",
+	"common-runtime/std"
 ]
 runtime-benchmarks = ["frame-benchmarking/runtime-benchmarks", "pallet-xcm/runtime-benchmarks"]
 try-runtime = ["frame-support/try-runtime"]
