+++ pallets/proposals/Cargo.toml
@@ -44,6 +44,7 @@ orml-xtokens = { git = "https://github.com/open-web3-stack/open-runtime-module-l
 # Local Dependencies
 common-traits = { path = "../../libs/common-traits", default-features = false }
 common-types = { path = "../../libs/common-types", default-features = false }
+pallet-deposits = { path = "../deposits", default-features = false }
 
 [dev-dependencies]
 orml-currencies = { git = "https://github.com/open-web3-stack/open-runtime-module-library", branch = "polkadot-v0.9.39"}
